refactor: split os_apps.rs into directory module (readability ratchet)
Split 515-line os_apps.rs into os_apps/mod.rs (247 lines) and
os_apps/tests.rs (274 lines) to satisfy the 500-line file limit.
